OT208-60
The endpoint of organization/public is modified to obtain a list of Slides of an organization sorted according to the Slide orders field. An important detail is that in the tests it was verified that when uploading an image in S3 of amazon to assign a url to a slide, a copy of the image was generated in the root directory of the project, so the myFile.delete() line is added after uploading the file and before returning the url.